Removed some autoloads and moved content type and direct url into seperate modules
